+++ subgraphs/base/src/price/PriceLookup.ts
@@ -1,17 +1,15 @@
 import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
 
 import { PriceHandler, PriceLookupResult } from "../../../shared/src/price/PriceHandler";
-import { PriceHandlerERC4626 } from "../../../shared/src/price/PriceHandlerERC4626";
 import { PriceHandlerUniswapV2 } from "../../../shared/src/price/PriceHandlerUniswapV2";
 import { PriceHandlerUniswapV3 } from "../../../shared/src/price/PriceHandlerUniswapV3";
 import { getUSDRate } from "../../../shared/src/price/PriceRouter";
 import {
   ERC20_OHM,
-  ERC20_USDS,
+  ERC20_USDC,
   ERC20_WETH,
-  ERC4626_SUSDS,
   LP_UNISWAP_V2_OHM_WETH,
-  LP_UNISWAP_V3_OHM_SUSDS,
+  LP_UNISWAP_V3_OHM_USDC,
   UNISWAP_V3_POSITION_MANAGER
 } from "../contracts/Constants";
 import { getContractName } from "../contracts/Contracts";
@@ -19,8 +17,7 @@ import { getBaseTokenRate, isBaseToken } from "./PriceBase";
 
 export const PRICE_HANDLERS: PriceHandler[] = [
   new PriceHandlerUniswapV2([ERC20_OHM, ERC20_WETH], LP_UNISWAP_V2_OHM_WETH, getContractName),
-  new PriceHandlerUniswapV3([ERC20_OHM, ERC4626_SUSDS], LP_UNISWAP_V3_OHM_SUSDS, UNISWAP_V3_POSITION_MANAGER, getContractName),
-  new PriceHandlerERC4626(ERC4626_SUSDS, ERC20_USDS, getContractName),
+  new PriceHandlerUniswapV3([ERC20_OHM, ERC20_USDC], LP_UNISWAP_V3_OHM_USDC, UNISWAP_V3_POSITION_MANAGER, getContractName),
 ];
 
 /**
@@ -70,10 +67,10 @@ export function getPrice(tokenAddress: string, block: BigInt): BigDecimal {
   const priceResult = getPriceRecursive(tokenAddress, block, null);
 
   if (priceResult === null) {
-    log.warning("Unable to determine price for token {} ({}) at block {}", [getContractName(tokenAddress), tokenAddress, block.toString()]);
+    log.warning("{}: Unable to determine price for token {} ({}) at block {}", [FUNC, getContractName(tokenAddress), tokenAddress, block.toString()]);
     return BigDecimal.zero();
   }
 
-  log.debug("Price for {} ({}) at block {} was: {}", [getContractName(tokenAddress), tokenAddress, block.toString(), priceResult.price.toString()]);
+  log.debug("{}: Price for {} ({}) at block {} was: {}", [FUNC, getContractName(tokenAddress), tokenAddress, block.toString(), priceResult.price.toString()]);
   return priceResult.price;
 }