+++ subgraphs/shared/tests/price/PriceHandlerUniswapV3.test.ts
@@ -81,6 +81,7 @@ const WALLET_ADDRESS = "0x18a390bD45bCc92652b9A91AD51Aed7f1c1358f5".toLowerCase(
 
 const ERC20_OHM = "0x64aa3364F17a4D01c6f1751Fd97C2BD3D7e7f1D5".toLowerCase();
 const ERC20_WETH = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2".toLowerCase();
+const ERC20_USDC = "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913".toLowerCase();
 
 const OHM_WETH_POOL = "0x88051b0eea095007d3bef21ab287be961f3d8598".toLowerCase();
 const OHM_WETH_SQRTPRICEX96 = BigInt.fromString("198259033222864761237442349019430");
@@ -89,6 +90,14 @@ const OHM_WETH_POSITION_ID = BigInt.fromString("222");
 const OHM_WETH_POSITION_TICK_LOWER: i32 = -887220;
 const OHM_WETH_POSITION_TICK_UPPER: i32 = 887220;
 const OHM_WETH_POSITION_LIQUIDITY = BigInt.fromString("346355586036686019");
+
+const OHM_USDC_POOL = "0x183ea22691c54806FE96555436dd312b6BeFAc2F".toLowerCase();
+const OHM_USDC_SQRTPRICEX96 = BigInt.fromString("11823183971744406029263508776");
+const OHM_USDC_TICK: i32 = -38048;
+const OHM_USDC_POSITION_ID = BigInt.fromString("1872809");
+const OHM_USDC_POSITION_TICK_LOWER: i32 = -44200;
+const OHM_USDC_POSITION_TICK_UPPER: i32 = 887220;
+const OHM_USDC_POSITION_LIQUIDITY = BigInt.fromString("11264485942092");
 
 export const mockFpisFraxPair = (): void => {
   mockRateUniswapV3(
@@ -115,6 +124,20 @@ export const mockOhmWethPair = (): void => {
     18,
     toBigInt(BigDecimal.fromString("139219.0068728")),
     toBigInt(BigDecimal.fromString("871.50742434")),
+  );
+};
+
+export const mockOhmUsdcPair = (): void => {
+  mockRateUniswapV3(
+    OHM_USDC_POOL,
+    OHM_USDC_SQRTPRICEX96,
+    OHM_USDC_TICK,
+    ERC20_OHM,
+    ERC20_USDC,
+    9,
+    6,
+    toBigInt(BigDecimal.fromString("75504.35396532")),
+    toBigInt(BigDecimal.fromString("445506.63314")),
   );
 };
 
@@ -153,6 +176,48 @@ export const mockOhmWethPosition = (): void => {
       ethereum.Value.fromI32(OHM_WETH_POSITION_TICK_LOWER), // tickLower
       ethereum.Value.fromI32(OHM_WETH_POSITION_TICK_UPPER), // tickUpper
       ethereum.Value.fromUnsignedBigInt(OHM_WETH_POSITION_LIQUIDITY), // liquidity
+      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // feeGrowthInside0X128
+      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // feeGrowthInside1X128
+      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // tokensOwed0
+      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // tokensOwed1
+    ]);
+};
+
+export const mockOhmUsdcPosition = (): void => {
+  // positionManager.balanceOf()
+  createMockedFunction(
+    Address.fromString(UNISWAP_V3_POSITION_MANAGER),
+    "balanceOf",
+    "balanceOf(address):(uint256)",
+  )
+    .withArgs([ethereum.Value.fromAddress(Address.fromString(WALLET_ADDRESS))])
+    .returns([ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(1))]);
+
+  // positionManager.tokenOfOwnerByIndex()
+  createMockedFunction(
+    Address.fromString(UNISWAP_V3_POSITION_MANAGER),
+    "tokenOfOwnerByIndex",
+    "tokenOfOwnerByIndex(address,uint256):(uint256)",
+  )
+    .withArgs([ethereum.Value.fromAddress(Address.fromString(WALLET_ADDRESS)), ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0))])
+    .returns([ethereum.Value.fromUnsignedBigInt(OHM_USDC_POSITION_ID)]);
+
+  // positionManager.positions()
+  createMockedFunction(
+    Address.fromString(UNISWAP_V3_POSITION_MANAGER),
+    "positions",
+    "positions(uint256):(uint96,address,address,address,uint24,int24,int24,uint128,uint256,uint256,uint128,uint128)",
+  )
+    .withArgs([ethereum.Value.fromUnsignedBigInt(OHM_USDC_POSITION_ID)])
+    .returns([
+      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // nonce
+      ethereum.Value.fromAddress(Address.zero()), // operator
+      ethereum.Value.fromAddress(Address.fromString(ERC20_OHM)), // token0
+      ethereum.Value.fromAddress(Address.fromString(ERC20_USDC)), // token1
+      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(10000)), // fee
+      ethereum.Value.fromI32(OHM_USDC_POSITION_TICK_LOWER), // tickLower
+      ethereum.Value.fromI32(OHM_USDC_POSITION_TICK_UPPER), // tickUpper
+      ethereum.Value.fromUnsignedBigInt(OHM_USDC_POSITION_LIQUIDITY), // liquidity
       ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // feeGrowthInside0X128
       ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // feeGrowthInside1X128
       ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(0)), // tokensOwed0
@@ -348,7 +413,28 @@ describe("getUnderlyingTokenBalance", () => {
     const ohmBalance = handler.getUnderlyingTokenBalance(WALLET_ADDRESS, ERC20_OHM, BLOCK);
     const wethBalance = handler.getUnderlyingTokenBalance(WALLET_ADDRESS, ERC20_WETH, BLOCK);
 
-    assert.stringEquals("138431.489223295", ohmBalance.toString());
-    assert.stringEquals("866.581676263788419538", wethBalance.toString());
+    assert.stringEquals("138410.423", ohmBalance.truncate(4).toString());
+    assert.stringEquals("866.7135", wethBalance.truncate(4).toString());
+  });
+
+  test("OHM-USDC balances are correct", () => {
+    const contractLookup: ContractNameLookup = (tokenAddress: string): string => {
+      if (addressesEqual(tokenAddress, ERC20_USDC)) {
+        return "USDC";
+      }
+
+      return "OHM";
+    };
+
+    mockOhmUsdcPair();
+    mockOhmUsdcPosition();
+
+    const handler = new PriceHandlerUniswapV3([ERC20_OHM, ERC20_USDC], OHM_USDC_POOL, UNISWAP_V3_POSITION_MANAGER, contractLookup);
+
+    const ohmBalance = handler.getUnderlyingTokenBalance(WALLET_ADDRESS, ERC20_OHM, BLOCK);
+    const usdcBalance = handler.getUnderlyingTokenBalance(WALLET_ADDRESS, ERC20_USDC, BLOCK);
+
+    assert.stringEquals("75484.2794", ohmBalance.truncate(4).toString());
+    assert.stringEquals("445136.3419", usdcBalance.truncate(4).toString());
   });
 });