+++ subgraphs/base/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Base Subgraph Changelog
 
+## 1.0.9 (2025-01-28)
+
+- Removes tracking of USDS, sUSDS and OHM-sUSDS LP
+- Adds tracking of USDC and OHM-USDC LP
+- Fixes issues with underlying token balances for Uniswap V3 positions
+
 ## 1.0.2 (2025-01-10)
 
 - Adds price handler for ERC4626 tokens